Fix focus bug when clicking on a token (#65)
Previously, the highlighted choice was getting cleared when a token was blurred, which was a silly thing to do since we blur the token and focus the component so we can better handle keyboard events. This would cause clicking on a token to highlight and then instantly unhighlight it.

Additionally, now when you start typing like normal while a token is highlighted, we focus the field, unhighlight the token, and input the text into the field so the user doesn't have to bother unhighlighting when typing.